+++ koala/cogs/announce/__init__.py
@@ -1,3 +1,3 @@
 from . import utils, db, log, models
-from .cog import Announce
+from .cog import Announce, setup
 from .announce_message import AnnounceMessage