+++ koala/cogs/verification/__init__.py
@@ -1,2 +1,2 @@
 from . import db, log, models
-from .cog import Verification
+from .cog import Verification, setup