+++ tests/cogs/twitch_alert/test_db.py
@@ -183,6 +183,18 @@ async def test_remove_team_from_ta(twitch_alert_db_manager_tables):
 
 
 @pytest.mark.asyncio()
+async def test_remove_team_from_ta_duplicate(twitch_alert_db_manager_tables):
+    test_add_team_to_ta_custom_message(twitch_alert_db_manager_tables, channel_id=590, guild_id=591)
+    test_add_team_to_ta_custom_message(twitch_alert_db_manager_tables, channel_id=590, guild_id=591)
+    await twitch_alert_db_manager_tables.remove_team_from_ta(590, "faze")
+
+    sql_select_team = select(TeamInTwitchAlert.custom_message)\
+        .where(and_(TeamInTwitchAlert.channel_id == 590, TeamInTwitchAlert.twitch_team_name == 'faze'))
+    with session_manager() as session:
+        assert session.execute(sql_select_team).one_or_none() is not None
+
+
+@pytest.mark.asyncio()
 async def test_remove_team_from_ta_invalid(twitch_alert_db_manager_tables):
     with pytest.raises(AttributeError,
                        match="Team name not found"):
@@ -291,3 +303,12 @@ async def test_delete_all_offline_streams_team(twitch_alert_db_manager_tables, b
         assert len(result) == 2
         assert result[0].message_id is None
         assert result[1].message_id is None
+
+
+@pytest.mark.asyncio
+async def test_create_alert_embed(twitch_alert_db_manager_tables):
+    stream_data = {'id': '3215560150671170227', 'user_id': '27446517',
+                   "user_name": "Monstercat", 'user_login': "monstercat", 'game_id': "26936", 'type': 'live',
+                   'title': 'Music 24/7'}
+
+    assert type(await twitch_alert_db_manager_tables.create_alert_embed(stream_data, None)) is discord.Embed