+++ CHANGELOG.md
@@ -3,6 +3,43 @@ All notable changes to KoalaBot will be documented in this file. A lot of these 
 administrators
 
 ## [Unreleased]
+
+## [0.5.8] - 28-04-2022
+### TwitchAlert
+- Fix user sql error
+
+## [0.5.7] - 28-04-2022
+### TwitchAlert
+- Fix issue where users alerts would sometimes delete themselves 
+
+## [0.5.6] - 19-04-2022
+### TwitchAlert
+- Fix issue where duplicate team alerts cannot be removed
+### Base
+- Command Checks now are presented well to the client
+
+## [0.5.5] - 19-04-2022
+### TwitchAlert
+- Fix issue where team twitch alerts would not show
+- Add logging and error handling to team loop
+
+## [0.5.4] - 12-04-2022
+### Base
+- Add scheduled activities for bot owners
+
+## [0.5.3] - 12-04-2022
+### TwitchAlert
+- Fix streams delete themselves when live (again)
+
+## [0.5.2] - 12-04-2022
+### TwitchAlert
+- Fix streams delete themselves when live
+
+## [0.5.1] - 11-04-2022
+### Vote
+- Fix error if someone leaves the server during a vote
+
+## [0.5.0] - 11-04-2022
 ### TwitchAlert
 - TwitchAlert completely reworked (part 1)
 - TwitchAlert commands have been altered (e.g. `k!twitchAddTeam` to `k!twitch addTeam`) see https://koalabot.uk/documentation for new syntax
@@ -250,6 +287,15 @@ manually if a timer is deleted.
 - `removeProtectedRoleColour <role_str>` Removes a role, via ID, mention or name, from the list of protected roles.
 
 [Unreleased]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.1.0...HEAD
+[0.5.8]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.7...v0.5.8
+[0.5.7]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.6...v0.5.7
+[0.5.6]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.5...v0.5.6
+[0.5.5]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.4...v0.5.5
+[0.5.4]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.3...v0.5.4
+[0.5.3]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.2...v0.5.3
+[0.5.2]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.1...v0.5.2
+[0.5.1]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.0...v0.5.1
+[0.5.0]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.4.5...v0.5.0
 [0.4.5]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.4.2...v0.4.3